+++ packages/core/solidity/src/contract.ts
@@ -10,11 +10,8 @@ export interface Contract {
   functions: ContractFunction[];
   constructorCode: string[];
   constructorArgs: FunctionArgument[];
-  constructorComments: string[];
   variables: string[];
-  shouldAutoTranspileImports: boolean;
-  shouldInstallContractsUpgradeable: boolean;
-  shouldUseUpgradesPluginsForProxyDeployment: boolean;
+  upgradeable: boolean;
 }
 
 export type Value = string | number | { lit: string } | { note: string; value: Value };
@@ -79,16 +76,12 @@ export interface NatspecTag {
 export class ContractBuilder implements Contract {
   readonly name: string;
   license: string = 'MIT';
-
-  shouldAutoTranspileImports: boolean = false;
-  shouldInstallContractsUpgradeable: boolean = false;
-  shouldUseUpgradesPluginsForProxyDeployment: boolean = false;
+  upgradeable = false;
 
   readonly natspecTags: NatspecTag[] = [];
 
   readonly constructorArgs: FunctionArgument[] = [];
   readonly constructorCode: string[] = [];
-  readonly constructorComments: string[] = [];
   readonly variableSet: Set<string> = new Set();
 
   private parentMap: Map<string, Parent> = new Map<string, Parent>();
@@ -207,15 +200,6 @@ export class ContractBuilder implements Contract {
 
   addConstructorCode(code: string) {
     this.constructorCode.push(code);
-  }
-
-  addConstructorComment(comment: string) {
-    if (this.shouldAutoTranspileImports) {
-      throw new Error(
-        'Constructor comments are not supported when `shouldAutoTranspileImports` is true, since constructor will be transformed into an initializer',
-      );
-    }
-    this.constructorComments.push(comment);
   }
 
   addFunctionCode(code: string, baseFn: BaseFunction, mutability?: FunctionMutability) {